requirements.txt
scipy==1.5.1
gym==0.10.9
numpy==1.19.4
einops==0.2.0
pandas==1.1.4
attr==0.3.1
quaternion==3.5.2.post4
tqdm==4.54.0
torch-scatter==2.0.6
